Add optional callback to Formbot#onChange
> ### Summary
I ran into a situation where I needed to call `onChange` for
field B that depended on the `onChange` of field A (see example) below.
I have also tried unsuccessfully to use the `onChange` prop on `Formbot`
to set a local state that I could then query. Adding this optional
callback seemed to be the most straightforward way of doing it but
other suggestions are always welcome.

```jsx
<Formbot>
  {({ values, onChange }) => (
    <Fragment>
      <Field>
        <Select
          value={values.status}
          onChange={value => {
            onChange('status', value);

            // `values` doesn't update fast enough to reflect the state change.
            if (values.status.length === 0) {
              onChange('doctor', '');
            }
          }}
        />
      </Field>

      <Field>
        <Select value={values.doctor} onChange={value => onChange('doctor', value)} />
      </Field>
    </Fragment>
  )}
</Formbot>
```